ib_tasks/interactors/create_or_update_task/field_response_validations/number_field_validation.py
from typing import Optional, List

from ib_tasks.exceptions.field_values_custom_exceptions import \
    InvalidNumberValue
from ib_tasks.interactors.create_or_update_task.field_response_validations. \
    base_field_validation import BaseFieldValidation
from ib_tasks.interactors.storage_interfaces.fields_dtos import \
    FieldWithGoFDisplayNameDTO


class NumberFieldValidationInteractor(BaseFieldValidation):

    def __init__(self, field_id: str, field_response: str):
        self.field_id = field_id
        self.field_response = field_response

    def validate_field_response(
            self,
            field_id_with_display_name_dtos: List[FieldWithGoFDisplayNameDTO]
    ) -> Optional[InvalidNumberValue]:
        invalid_number_value = not self.field_response.isdigit()
        if invalid_number_value:
            field_display_name = self.get_field_display_name(
                self.field_id, field_id_with_display_name_dtos)
            raise InvalidNumberValue(field_display_name, self.field_response)
        return
